refactor: reorganize frontend api layer
