--- front/src/api/log.jsx
@@ -1,3 +1,0 @@
-export const logError = (message, error, context) => {
-  console.error(`[Error] ${message}` + (context ? ` | Context: ${context}` : ''), error);
-};
